Extracted modal dialog from Message component

+++ frontend/src/components/Message.tsx
@@ -6,17 +6,10 @@ import {
   ListItemSecondaryAction,
   ListItemText,
   ListItemAvatar,
-  Avatar,
-  Dialog,
-  DialogTitle,
-  TextField,
-  DialogContent,
-  DialogActions,
-  Button,
-  DialogContentText
+  Avatar
 } from "@material-ui/core";
 import { DeleteOutline, EditOutlined } from "@material-ui/icons";
-import React, { useState } from "react";
+import React from "react";
 import {
   DELETE_MESSAGE,
   DELETE_MESSAGE_FROM_CACHE,
@@ -31,6 +24,7 @@ import {
   EditMessageMutation,
   EditMessageMutationVariables
 } from "../typescript/codegen";
+import { useModal } from "../contexts/ModalContext";
 
 interface Props {
   message: MessageListMessageFragment;
@@ -51,32 +45,37 @@ const Message: React.FC<Props> = ({ message, isOwner }) => {
     EditMessageMutationVariables
   >(EDIT_MESSAGE);
 
-  const [editMode, setEditMode] = useState(false);
-  const [deleteConfirmOpen, setDeleteConfirmOpen] = useState(false);
-  const [messageTextForEditing, setMessageTextForEditing] = useState(
-    message.messageText
-  );
+  const { confirm, form } = useModal();
 
   const deleteMessage = async () => {
-    const deleteMessageResult = await deleteMessageFromServer();
     if (
-      deleteMessageResult.data &&
-      deleteMessageResult.data.deleteMessage.success
+      await confirm({
+        title: "Delete Message?",
+        description:
+          "Deleting a message is permanent. Are you sure you want to do this?"
+      })
     ) {
-      deleteMessageFromCache();
+      const deleteMessageResult = await deleteMessageFromServer();
+      if (
+        deleteMessageResult.data &&
+        deleteMessageResult.data.deleteMessage.success
+      ) {
+        deleteMessageFromCache();
+      }
     }
   };
 
-  const toggleEdit = () => {
-    setEditMode(current => !current);
-  };
-
   const editMessage = () => {
-    toggleEdit();
-    editMessageMutation({
-      variables: {
-        messageId: message.id,
-        updatedText: messageTextForEditing
+    form({
+      title: "Edit Message",
+      initialValue: message.messageText,
+      onConfirm: formValue => {
+        editMessageMutation({
+          variables: {
+            messageId: message.id,
+            updatedText: formValue
+          }
+        });
       }
     });
   };
@@ -98,7 +97,7 @@ const Message: React.FC<Props> = ({ message, isOwner }) => {
                 color="primary"
                 edge="end"
                 aria-label="edit"
-                onClick={toggleEdit}
+                onClick={editMessage}
               >
                 <EditOutlined />
               </IconButton>
@@ -106,72 +105,11 @@ const Message: React.FC<Props> = ({ message, isOwner }) => {
                 color="primary"
                 edge="end"
                 aria-label="delete"
-                onClick={() => setDeleteConfirmOpen(true)}
+                onClick={deleteMessage}
               >
                 <DeleteOutline />
               </IconButton>
             </ListItemSecondaryAction>
-            <Dialog open={editMode} onClose={toggleEdit}>
-              <DialogTitle>Edit Message</DialogTitle>
-              <DialogContent>
-                <TextField
-                  autoFocus
-                  fullWidth
-                  value={messageTextForEditing}
-                  onChange={({ target }) =>
-                    setMessageTextForEditing(target.value)
-                  }
-                  onKeyPress={({ key }) => {
-                    // OnKeyDown and OnKeyUp doesn't toggle the dialog properly
-                    if (key === "Enter") {
-                      editMessage();
-                    }
-                  }}
-                />
-              </DialogContent>
-              <DialogActions>
-                <Button onClick={toggleEdit} color="primary">
-                  Cancel
-                </Button>
-                <Button
-                  color="primary"
-                  onClick={() => {
-                    editMessage();
-                  }}
-                >
-                  Submit
-                </Button>
-              </DialogActions>
-            </Dialog>
-            <Dialog
-              open={deleteConfirmOpen}
-              onClose={() => setDeleteConfirmOpen(false)}
-            >
-              <DialogTitle>Delete message?</DialogTitle>
-              <DialogContent>
-                <DialogContentText>
-                  Deleting a message is permanent. Are you sure you want to do
-                  this?
-                </DialogContentText>
-              </DialogContent>
-              <DialogActions>
-                <Button
-                  onClick={() => setDeleteConfirmOpen(false)}
-                  color="primary"
-                >
-                  Cancel
-                </Button>
-                <Button
-                  onClick={() => {
-                    deleteMessage();
-                    setDeleteConfirmOpen(false);
-                  }}
-                  color="primary"
-                >
-                  Confirm
-                </Button>
-              </DialogActions>
-            </Dialog>
           </>
         )}
       </ListItem>
